Move utility functions into util/ classes
