+++ src/util/platform.ts
