--- src/windowUtils.ts
@@ -1,25 +1,0 @@
-/*-----------------------------------------------------------------------------------------------
- *  Copyright (c) Red Hat, Inc. All rights reserved.
- *  Licensed under the MIT License. See LICENSE file in the project root for license information.
- *-----------------------------------------------------------------------------------------------*/
-
-import { window, Terminal, TerminalOptions } from 'vscode';
-import * as path from 'path';
-import { Platform } from './platform';
-
-const toolsLocation = path.resolve(Platform.getUserHomePath(), '.vs-openshift');
-
-export function createTerminal(name: string, cwd: string, env: NodeJS.ProcessEnv = process.env): Terminal {
-    const finalEnv: NodeJS.ProcessEnv = {};
-    Object.assign(finalEnv, env);
-    const key = process.platform === 'win32' ? 'Path' : 'PATH';
-    finalEnv[key] = `${toolsLocation}${path.delimiter}${env[key]}`;
-
-    const options: TerminalOptions = {
-        cwd: cwd,
-        name: name,
-        env: finalEnv,
-        shellPath: process.platform === 'win32'? undefined : '/bin/bash'
-    };
-    return window.createTerminal(options);
-}
